got rid off "denite" plugin
it stopped working for no reason some time ago, Denite window opens but
there's no field to input to, it renders to a window and after that it's
just a read-only buffer of text, i can see the results but can't do
nothing with them but copying text.

i tried a lot of ways to make it work again, downgrade, update python
package for neovim, i'm tired of this, have no time to find a way to
reproduce with minimal vimrc to create an issue.

i just replaced few command with kinda native ones
(which don't use any plugin).