Disable “textwidth” limit
It’s often just stands in the way, and I have to turn it off right after opening
a file or a shell command to edit so that it doesn’t mess with the long lines.
It’s better to have it turned off by default and enabled for some specific
projects where it’s really needed.